Extracted FileBackupManager to its own file
