+++ src/csv_to_confluence.py
@@ -36,6 +36,7 @@ from packaging import version
 
 from common import Main, CSVHandler, ConfluenceConnection, ConfluenceNodeMapper, ErrorCSVHandler, InfoCSVHandler, ResourceLoader, SingletonABCMeta, \
     SingletonMeta, TimestampHandler
+from src.FileBackupManager import FileBackupManager
 
 
 class TemplatePageLoader(ResourceLoader):
@@ -731,37 +732,6 @@ class ConfluencePageHandler(ConfluenceHandler):
         for content_writer in self.__content_writers:
             template = content_writer.add_content_to_template_page(template, node_id)
         return template
-
-
-class FileBackupManager(ConfluenceHandler):
-    """
-    Backs up all files of the corresponding broker node ID on its confluence page.
-    Identically named attachments are overwritten when uploaded to Confluence.
-    """
-
-    def __init__(self):
-        super().__init__()
-        self.__working_dir = os.getenv('DIR.WORKING')
-
-    def backup_files(self, node_id: str):
-        """
-        Backs up files of the specified broker node ID by uploading them as attachments to the Confluence page.
-        """
-        self.__backup_files_with_line_ending(node_id, 'csv')
-        self.__backup_files_with_line_ending(node_id, 'txt')
-        self.__backup_files_with_line_ending(node_id, 'log')
-
-    def __backup_files_with_line_ending(self, node_id: str, line_ending: str):
-        node_dir = os.path.join(self.__working_dir, node_id)
-        files_list = self.__get_all_files_in_directory_with_line_ending(node_dir, line_ending)
-        name = self._mapper.get_node_value_from_mapping_dict(node_id, 'COMMON_NAME')
-        for filename in files_list:
-            filepath = os.path.join(node_dir, filename)
-            self._confluence.upload_file_as_attachement_to_page(name, filepath)
-
-    @staticmethod
-    def __get_all_files_in_directory_with_line_ending(directory: str, line_ending: str) -> list:
-        return [name_file for name_file in os.listdir(directory) if name_file.endswith(line_ending)]
 
 
 class SummaryTableCreator:
